+++ CHANGELOG.md
@@ -71,6 +71,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Missing note in Core::print_basic_help() about how to exit OpenRPG.
 - Fixed a bug in the ExpressionTree where the `globalReadOffset` was not being reset when a new expression was set
 - Fixed a bug when stringing multiple rolls together that would causing the tree to parse an expression such as 2d6+4d8 as (2d6+4)d8
+- Fixed a bug in NameGenerator that wasn't resetting the raceFile when a new race was set.
 
 ### Removed
 - Obliterated the old Name Generator test src/tests/name-generator-test.cpp